creation des petits fichier dans le dossier parser pour faire preparsing.c plus propre; ajout parser dans le main de readline.c
