+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+	"files.associations": {
+		"minishell.h": "c",
+		"limits.h": "c"
+	}
+}
