+++ Makefile
@@ -15,7 +15,11 @@ NAME		= minishell
 SRCS		= readline.c utils.c builtin/utils.c builtin/cd.c builtin/echo.c \
 				builtin/env.c builtin/exit.c builtin/export.c builtin/pwd.c \
 				builtin/unset.c binaries/main_bin.c binaries/path_utils.c \
-				exit/exit.c env/env_init.c pipe/pipe.c pipe/pipe_utils.c
+				exit/exit.c env/env_init.c pipe/pipe.c pipe/pipe_utils.c \
+				parsing/lexer/ft_split_tokens.c parsing/lexer/nbr_tokens.c parsing/lexer/which_type_is_it.c \
+				parsing/parser/preparsing.c parsing/parser/tools_parser.c parsing/parser/fill_args_parser.c \
+				parsing/parser/fill_redir_parser.c parsing/parser/verif_parsing.c\
+				libft/ft_strjoin.c libft/ft_split.c
 
 OBJS		= $(SRCS:.c=.o)
