+++ exit/exit.c
@@ -49,7 +49,7 @@ int ft_proper_free(t_simple_command *c_table)
     if (c_table->args)
         ft_free_2dstr(c_table->args);
     if (c_table->redir)
-        ft_free_redir(c_table->redir);
+        ft_free_redir(&(c_table->redir));
     if (c_table->infile >= 0)
         close(c_table->infile);
     if (c_table->outfile >= 0)
