+++ parsing/lexer/which_type_is_it.c
@@ -10,7 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "lexer.h"
 #include "../../minishell.h"
